requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
cloudinary==1.38.0
imageio-ffmpeg==0.4.9
imageio==2.34.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
psutil==5.9.8